server/requirements.txt
fastapi
websockets
uvicorn
paho-mqtt